+++ comics/comics/stickydillybuns.py
@@ -11,11 +11,15 @@ class ComicData(ComicDataBase):
 
 
 class Crawler(CrawlerBase):
-    history_capable_date = '2013-01-07'
+    history_capable_days = 50
     schedule = 'Mo,Fr'
     time_zone = 'US/Eastern'
 
     def crawl(self, pub_date):
-        url = 'http://www.stickydillybuns.com/comics/sdb%s.png' % (
-            pub_date.strftime('%Y%m%d'))
-        return CrawlerImage(url)
+        feed = self.parse_feed('http://www.stickydillybuns.com/comic.rss')
+        for entry in feed.for_date(pub_date):
+            title = entry.title.replace('Sticky Dilly Buns - ', '')
+            page_url = entry.link.replace(' ', '+')
+            page = self.parse_page(page_url)
+            url = page.src('#comic img')
+            return CrawlerImage(url, title)